Use Tcl_GetHashKey() to use a hash entry key
The Tcl_HashEntry structure is private to Tcl and its fields should not
be directly accessed outside Tcl. Instead, Tcl_{Get,Set}Hash{Key,Value}
macros should be used.

Tk is probably an exception here as it's maintained together with the
Tcl Core but it would still be better to use a well-defined interface
that is unlikely to change. While the Tcl_HashEntry.tablePtr field
should not be directly accessed as well, it allows us not to change
the Tk API to pass hash tables to use the Tcl_{Get,Set}Hash{Key,Value}
macros. We can discuss with upstream if it should be changed as well.

This change was needed as Tcl_HashEntry.key.string was renamed to bytes
in the proposed CheriABI patch [0].

Discussed with:	@jrtc27

[0] https://core.tcl-lang.org/tcl/tktview?name=37108037b9